requirements_test.txt
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-homeassistant-custom-component==0.13.205
respx==0.21.1
